FIX-1968 move account menu to settings
